MobileNetworkSettings: Check CarrierConfig on isWorldMode
Add support for checking the CarrierConfig version of the World
Phone config. If this is not present, fall back to the old
deprecated way of looking it up from the Telephony settings.

Not having this prevented the world phone config from working if
you fully migrated to CarrierConfig - you had to have both versions
or you would fail this check

Change-Id: I73939f1055849793d108b02bf034ab79e360c4e4